requirements.txt
Django
djangorestframework
google-api-python-client
oauth2client
python-social-auth
psycopg2